add: 정렬, 키워드 필터 추가
